deeplab

+++ 1611.08408/reader.py
@@ -35,7 +35,7 @@ class Reader():
             images[:, :, :, i] = img
 
             gt = cv2.imread(os.path.join(self.data_root, 'masks', self.filelist[self.index + i] + '.png'),
-                            cv2.IMREAD_GRAYSCALE).astype(float)
+                            cv2.IMREAD_GRAYSCALE)
             gt_interp = cv2.resize(gt, ((img.shape[0]-1)/8+1, (img.shape[1]-1)/8+1), cv2.INTER_LINEAR)
             ground_truths_interp[:, :, i] = gt_interp
